Refactor: Seperating Game Protocols in multiple files fix: Fixed wrong logic on deleting Discord Messages
