+++ src/discovery/__init__.py
@@ -2,10 +2,7 @@
 Server discovery and game wrapper package
 """
 
-from .network_scanner import NetworkScanner, DiscoveryEngine, ServerResponse, BroadcastProtocol, RenegadeXBroadcastProtocol
+from .network_scanner import NetworkScanner, DiscoveryEngine
+from .protocols import ServerResponse
 
-__all__ = ['NetworkScanner', 'DiscoveryEngine', 'ServerResponse', 'BroadcastProtocol', 'RenegadeXBroadcastProtocol']
-
-"""
-Discovery package for game server scanning and monitoring.
-""" 
+__all__ = ['NetworkScanner', 'DiscoveryEngine', 'ServerResponse']
