+++ src/discovery/protocols/warcraft3.py
@@ -1,11 +1,13 @@
 from dataclasses import dataclass
-from typing import Optional, Dict, Any, Tuple
+from typing import Optional, Dict, Any, Tuple, List
 import socket
 import struct
 import asyncio
 import logging
-from discovery.protocol_base import ProtocolBase
+import ipaddress
+from ..protocol_base import ProtocolBase
 from opengsq.binary_reader import BinaryReader
+from .common import ServerResponse
 
 # Setup logger
 logger = logging.getLogger(__name__)
@@ -171,6 +173,199 @@ class Warcraft3Protocol(ProtocolBase):
                 logger.error(f"Error in broadcast listener: {e}")
                 continue
     
+    async def scan_servers(self, scan_ranges: List[str]) -> List[ServerResponse]:
+        """
+        Scan for Warcraft 3 servers using broadcast queries.
+        
+        Args:
+            scan_ranges: List of network ranges to scan
+            
+        Returns:
+            List of ServerResponse objects for Warcraft3 servers
+        """
+        servers = []
+        
+        # Default game version to use for queries (Warcraft 3 TFT)
+        default_game_version = 0x1A  # Version 1.26a
+        
+        # For each network range, send broadcast queries
+        for network_range in scan_ranges:
+            try:
+                network = ipaddress.ip_network(network_range, strict=False)
+                broadcast_addr = str(network.broadcast_address)
+                
+                logger.debug(f"Broadcasting Warcraft3 query to {broadcast_addr}:{self.WARCRAFT3_PORT}")
+                
+                # Send broadcast query and collect responses
+                responses = await self._send_broadcast_query(
+                    broadcast_addr, self.WARCRAFT3_PORT, default_game_version
+                )
+                
+                # Process responses
+                for response_data, sender_addr in responses:
+                    try:
+                        server_info = await self._parse_warcraft3_response(response_data)
+                        if server_info:
+                            server_response = ServerResponse(
+                                ip_address=sender_addr[0],
+                                port=sender_addr[1],
+                                game_type='warcraft3',
+                                server_info=server_info,
+                                response_time=0.0
+                            )
+                            servers.append(server_response)
+                            logger.debug(f"Discovered Warcraft3 server: {sender_addr[0]}:{sender_addr[1]}")
+                        
+                    except Exception as e:
+                        logger.debug(f"Failed to parse Warcraft3 response from {sender_addr}: {e}")
+                        
+            except Exception as e:
+                logger.error(f"Error broadcasting to network {network_range}: {e}")
+        
+        return servers
+    
+    async def _send_broadcast_query(self, broadcast_addr: str, port: int, game_version: int) -> List[Tuple[bytes, Tuple[str, int]]]:
+        """
+        Send a Warcraft 3 broadcast query and collect responses.
+        
+        Args:
+            broadcast_addr: Broadcast address to send to
+            port: Port to send to
+            game_version: Game version to use in query
+            
+        Returns:
+            List of tuples containing (response_data, sender_address)
+        """
+        responses = []
+        
+        try:
+            # Create UDP socket for sending and receiving
+            sock = socket.socket(socket.AF_INET, socket.SOCK_DGRAM)
+            sock.setsockopt(socket.SOL_SOCKET, socket.SO_BROADCAST, 1)
+            sock.setsockopt(socket.SOL_SOCKET, socket.SO_REUSEADDR, 1)
+            sock.setblocking(False)
+            
+            # Create search game packet
+            search_packet = self.create_search_game_packet(game_version)
+            
+            # Send broadcast query
+            await asyncio.get_event_loop().sock_sendto(sock, search_packet, (broadcast_addr, port))
+            logger.debug(f"Sent Warcraft3 broadcast query to {broadcast_addr}:{port}")
+            
+            # Listen for responses
+            loop = asyncio.get_event_loop()
+            end_time = loop.time() + self.timeout
+            
+            while loop.time() < end_time:
+                try:
+                    remaining_time = end_time - loop.time()
+                    if remaining_time <= 0:
+                        break
+                    
+                    # Wait for response with timeout
+                    data, addr = await asyncio.wait_for(
+                        loop.sock_recvfrom(sock, 2048),
+                        timeout=min(remaining_time, 1.0)
+                    )
+                    
+                    logger.debug(f"Received {len(data)} bytes from {addr[0]}:{addr[1]}")
+                    responses.append((data, addr))
+                    
+                except asyncio.TimeoutError:
+                    continue
+                except Exception as e:
+                    logger.debug(f"Error receiving response: {e}")
+                    continue
+            
+            sock.close()
+            
+        except Exception as e:
+            logger.error(f"Error sending Warcraft3 broadcast query: {e}")
+        
+        return responses
+
+    async def _parse_warcraft3_response(self, response_data: bytes) -> Optional[Dict[str, Any]]:
+        """
+        Parse a Warcraft 3 server response.
+        
+        Args:
+            response_data: Raw response data from Warcraft3 server
+            
+        Returns:
+            Dictionary containing parsed server information, or None if parsing failed
+        """
+        try:
+            # Check if this is a valid Warcraft3 response
+            if len(response_data) < 4:
+                return None
+            
+            # Check for Warcraft3 protocol signature (0xF7)
+            if response_data[0] != 0xF7:
+                return None
+            
+            # Check for game info response (0x30)
+            if response_data[1] != 0x30:
+                return None
+            
+            # Create a BinaryReader for parsing
+            br = BinaryReader(response_data)
+            
+            # Skip protocol signature and packet type
+            br.read_bytes(2)
+            
+            # Read packet size
+            packet_size = int.from_bytes(br.read_bytes(2), 'little')
+            
+            # Read game version info
+            product = br.read_bytes(4).decode('ascii', errors='ignore')
+            version = int.from_bytes(br.read_bytes(4), 'little')
+            host_counter = int.from_bytes(br.read_bytes(4), 'little')
+            entry_key = int.from_bytes(br.read_bytes(4), 'little')
+            
+            # Read game name (null-terminated)
+            game_name = ""
+            while br.remaining_bytes() > 0:
+                char = int.from_bytes(br.read_bytes(1), 'little')
+                if char == 0:
+                    break
+                game_name += chr(char)
+            
+            # Skip unknown byte if available
+            if br.remaining_bytes() > 0:
+                br.read_bytes(1)
+            
+            # Read remaining fields if available
+            slots_total = 0
+            slots_used = 0
+            if br.remaining_bytes() >= 12:
+                # Skip settings string (find next null terminator)
+                while br.remaining_bytes() > 0:
+                    char = int.from_bytes(br.read_bytes(1), 'little')
+                    if char == 0:
+                        break
+                
+                if br.remaining_bytes() >= 12:
+                    slots_total = int.from_bytes(br.read_bytes(4), 'little')
+                    game_flags = int.from_bytes(br.read_bytes(4), 'little')
+                    slots_used = int.from_bytes(br.read_bytes(4), 'little')
+            
+            return {
+                'name': game_name or f"Warcraft3 Server",
+                'map': "Unknown Map",
+                'product': product,
+                'version': version,
+                'players': slots_used,
+                'max_players': slots_total,
+                'game': 'Warcraft III',
+                'host_counter': host_counter,
+                'entry_key': entry_key
+            }
+            
+        except Exception as e:
+            logger.debug(f"Failed to parse Warcraft3 response: {e}")
+        
+        return None
+
     async def query_server(self, address: str, port: int, game_version: int) -> Optional[Dict[str, Any]]:
         """
         Query a specific Warcraft 3 server for details.
